Update to include processing history.

+++ includes/editPinkieExpenses.php
@@ -4,6 +4,11 @@ secureSessionStart();
 
 if(strcmp($_POST['mode'], "delete") == 0) // We are in delete mode.
 {
+  if(strlen($_POST['fundID']) == 0)
+  {
+    // No fundID was set.
+    return;
+  }
   $_db = getMysqli();
   $_sql = "DELETE FROM Expenses WHERE ExpenseID=?";
   $_stmt = $_db->prepare((string)$_sql);
@@ -13,12 +18,13 @@ if(strcmp($_POST['mode'], "delete") == 0) // We are in delete mode.
 
   if ($_stmt->errno)
   {
-    onErrorInternal("editPinkieExpenses::Delete()", $_stmt->error);
+    onErrorInternal("editPinkieExpenses::editDelete()", $_stmt->error);
   }
   $_stmt->close();
   // Close up the database connection.
   $_db->close();
   echo "OKAY";
+  logGeneral($_POST['pinkieID'], $_SESSION['Username'], "Expense was deleted by: ".getName());
   return;
 }
 if(strcmp($_POST['mode'], "edit") == 0) // We are in edit mode.
@@ -44,18 +50,19 @@ if(strcmp($_POST['mode'], "edit") == 0) // We are in edit mode.
   // Close up the database connection.
   $_db->close();
   echo "OKAY";
+  logGeneral($_POST['pinkieID'], $_SESSION['Username'], "Expense was edited by: ".getName());
   return;
 
 }
 if(strcmp($_POST['mode'], "add") == 0) // We are in add mode.
 {
 
-  if(strlen($_POST['fundID']) == 0)
+  if(strlen($_POST['pinkieId']) == 0)
   {
-    // No fundID was set.
+    // No pinkieId was set.
     return;
   }
-  
+
   $_db = getMysqli();
   $_sql = "INSERT INTO Expenses (PinkieID, Amount, FundID) Values(?,?,?)";
   $_stmt = $_db->prepare((string)$_sql);
@@ -65,12 +72,13 @@ if(strcmp($_POST['mode'], "add") == 0) // We are in add mode.
 
   if ($_stmt->errno)
   {
-    onErrorInternal("editPinkieExpenses::editUpdate()", $_stmt->error);
+    onErrorInternal("editPinkieExpenses::editAdd()", $_stmt->error);
   }
   $_stmt->close();
   // Close up the database connection.
   $_db->close();
   echo "OKAY";
+  logGeneral($_POST['pinkieID'], $_SESSION['Username'], "Expense was added by: ".getName());
   return;
 }
